hot fix

+++ file_io/3-cp.c
@@ -88,11 +88,11 @@ int main(int argc, char *argv[])
 
         check_args(argc);
         fd_from = open(argv[1], O_RDONLY);
-        check_file_from(fd_from, argv[1], -1, -1);
+        check_file_from((ssize_t)fd_from, argv[1], -1, -1);
 
         file_perm = S_IRUSR | S_IWUSR | S_IRGRP | S_IWGRP | S_IROTH;
         fd_to = open(argv[2], O_WRONLY | O_CREAT | O_TRUNC, file_perm);
-        check_file_to(fd_to, argv[2], fd_from, -1);
+        check_file_to((ssize_t)fd_to, argv[2], fd_from, -1);
 
         len_read = 1024;
 
@@ -114,4 +114,3 @@ int main(int argc, char *argv[])
 
         return (0);
 }
-
